+++ packages/lib/package/components/MazInputPrice.vue
@@ -3,10 +3,29 @@
     v-model="displayPrice"
     class="maz-input-price"
     v-bind="$attrs"
-    left-icon="cash"
     @focus="isActive = true"
     @blur="isActive = false"
-  />
+  >
+    <template #left-icon>
+      <slot v-if="!noIcon" name="left-icon">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke="currentColor"
+          aria-hidden="true"
+          class="maz-text-mute maz-ml-1 maz-h-6 maz-w-6"
+        >
+          <path
+            stroke-linecap="round"
+            stroke-linejoin="round"
+            stroke-width="2"
+            d="M17 9V7a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2m2 4h10a2 2 0 002-2v-6a2 2 0 00-2-2H9a2 2 0 00-2 2v6a2 2 0 002 2zm7-5a2 2 0 11-4 0 2 2 0 014 0z"
+          />
+        </svg>
+      </slot>
+    </template>
+  </MazInput>
 </template>
 
 <script lang="ts" setup>
@@ -20,6 +39,7 @@
     locale: { type: String, default: 'fr-FR' },
     min: { type: Number, default: 0 },
     max: { type: Number, default: Infinity },
+    noIcon: { type: Boolean, default: false },
   })
 
   const emits = defineEmits(['update:model-value', 'formatted'])
@@ -61,11 +81,11 @@
   const displayPrice = computed({
     get: () => {
       if (isActive.value) return valueString.value
-      if (props.modelValue) return priceFormatted.value
+      if (typeof props.modelValue === 'number') return priceFormatted.value
       return undefined
     },
     set: (value) => {
-      if (!value) {
+      if (Number.isNaN(value)) {
         emitValues(undefined)
       } else {
         const adjustedPrice = getAdjustedPrice(value)
@@ -75,7 +95,8 @@
   })
 
   const emitValues = async (newValue?: number) => {
-    const adjustedPrice = newValue ? getAdjustedPrice(newValue) : undefined
+    const adjustedPrice =
+      typeof newValue === 'number' ? getAdjustedPrice(newValue) : undefined
     emits('update:model-value', adjustedPrice)
 
     await nextTick()